last refactoring before migrating to the project to typescript
