+++ client/src/components/BookingWidget.jsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { differenceInCalendarDays } from "date-fns";
 import axios from "axios";
 import { Navigate } from "react-router-dom";
-import { UserContext } from "./UserContext";
+import { UserContext } from "../UserContext";
 
 export default function BookingWidget({place}) {
     const [checkIn,setCheckIn] = useState('');
@@ -84,6 +84,5 @@ export default function BookingWidget({place}) {
             )}
         </button>
     </div>
-
     );
 }
