+++ client/src/components/Image.jsx
