+++ client/src/components/PlaceGallery.jsx
@@ -1,5 +1,7 @@
 import { useState } from "react";
 import Image from "./Image";
+import CloseIcon from "./icons/CloseIcon";
+import ShowMorePhotoIcon from "./icons/ShowMorePhotoIcon";
 
 export default function PlaceGallery({place}) {
     const [showAllPhotos,setShowAllPhotos] = useState(false);
@@ -11,9 +13,7 @@ export default function PlaceGallery({place}) {
                     <div>
                         <h2 className="text-3xl mr-48">Photos of {place.title}</h2>
                         <button onClick={() => setShowAllPhotos(false)} className="fixed right-12 top-8 flex gap-1 py-2 px-4 rounded-2xl shadow shadow-black bg-white text-black">
-                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="w-6 h-6">
-                            <path fillRule="evenodd" d="M5.47 5.47a.75.75 0 0 1 1.06 0L12 10.94l5.47-5.47a.75.75 0 1 1 1.06 1.06L13.06 12l5.47 5.47a.75.75 0 1 1-1.06 1.06L12 13.06l-5.47 5.47a.75.75 0 0 1-1.06-1.06L10.94 12 5.47 6.53a.75.75 0 0 1 0-1.06Z" clipRule="evenodd" />
-                        </svg>
+                        <CloseIcon />
                             Close photos
                         </button>
                     </div>
@@ -49,9 +49,7 @@ export default function PlaceGallery({place}) {
             </div>
         </div>
         <button onClick={() => setShowAllPhotos(true)} className="flex gap-1 absolute bottom-2 right-2 py-2 px-4 bg-white rounded-2xl shadow shadow-md shadow-gray-500">
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="w-6 h-6">
-            <path fillRule="evenodd" d="M1.5 6a2.25 2.25 0 0 1 2.25-2.25h16.5A2.25 2.25 0 0 1 22.5 6v12a2.25 2.25 0 0 1-2.25 2.25H3.75A2.25 2.25 0 0 1 1.5 18V6ZM3 16.06V18c0 .414.336.75.75.75h16.5A.75.75 0 0 0 21 18v-1.94l-2.69-2.689a1.5 1.5 0 0 0-2.12 0l-.88.879.97.97a.75.75 0 1 1-1.06 1.06l-5.16-5.159a1.5 1.5 0 0 0-2.12 0L3 16.061Zm10.125-7.81a1.125 1.125 0 1 1 2.25 0 1.125 1.125 0 0 1-2.25 0Z" clipRule="evenodd" />
-        </svg>
+        <ShowMorePhotoIcon />
             Show more photos
         </button>
     </div>
