+++ client/src/components/PlaceImg.jsx
