+++ client/src/components/UserPlaceItem.jsx
@@ -1,0 +1,16 @@
+import PlaceImg from "./PlaceImg";
+import { Link } from "react-router-dom";
+
+export default function UserPlaceItem({place}) {
+  return (
+    <Link to={'/account/places/'+place._id} className="flex cursor-pointer  gap-4 bg-gray-100 p-4 rounded-2xl">
+        <div className="flex w-32 h-32 bg-gray-300 grow shrink-0">
+            <PlaceImg place={place} />
+        </div>
+        <div className="grow-0 shrink">
+            <h2 className="text-xl">{place.title}</h2>
+            <p className="text-sm mt-2">{place.description}</p>
+        </div>
+    </Link>  
+    )
+}
