+++ client/src/pages/IndexPage.jsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from "react";
 import axios from "axios";
-import { Link } from "react-router-dom";
-import Image from "../Image";
+import PlaceItem from "../components/PlaceItem";
 
 export default function IndexPage() {
     const [places,setPlaces] = useState([]);
@@ -10,21 +9,11 @@ export default function IndexPage() {
             setPlaces(response.data);
         });
     }, []);
+    
     return(
         <div className="mt-8 grid gap-6 gap-x-6 gap-y-8 grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
             {places.length > 0 && places.map(place => (
-                <Link to={'/place/'+place._id} key={place._id}>
-                    <div className="bg-gray-500 mb-2 rounded-2xl flex">
-                    {place.photos?.[0] && (
-                        <Image className="rounded-2xl object-cover aspect-square" src={place.photos?.[0]} alt="" />
-                    )}
-                    </div>
-                    <h2 className="font-bold">{place.address}</h2>
-                    <h3 className="text-sm text-gray-500">{place.title}</h3>
-                    <div className="mt-2">
-                        <span className="font-bold">${place.price}</span> per night
-                    </div>
-                </Link>
+                <PlaceItem place={place} key={place._id}/>
             ))}
         </div>
     )
