+++ client/src/pages/PlacesPage.jsx
@@ -2,7 +2,8 @@ import { Link } from "react-router-dom";
 import AccountNav from "../components/AccountNav";
 import { useEffect, useState } from "react";
 import axios from "axios";
-import PlaceImg from "../PlaceImg";
+import UserPlaceItem from "../components/UserPlaceItem";
+import AddIcon from "../components/icons/AddIcon";
 
 export default function PlacesPage() {
     const [places,setPlaces] = useState([]);
@@ -18,23 +19,13 @@ export default function PlacesPage() {
             <AccountNav />
                 <div className="text-center">
                     <Link className="inline-flex gap-1 bg-primary text-white py-2 px-6 rounded-full" to={'/account/places/new'} >
-                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
-                            <path strokeLinecap="round" strokeLinejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
-                        </svg>
+                        <AddIcon />
                         Add new place
                     </Link>
                 </div>
                 <div className="mt-4">
                     {places.length > 0 && places.map(place => (
-                        <Link key={place._id} to={'/account/places/'+place._id} className="flex cursor-pointer  gap-4 bg-gray-100 p-4 rounded-2xl">
-                            <div className="flex w-32 h-32 bg-gray-300 grow shrink-0">
-                                <PlaceImg place={place} />
-                            </div>
-                            <div className="grow-0 shrink">
-                                <h2 className="text-xl">{place.title}</h2>
-                                <p className="text-sm mt-2">{place.description}</p>
-                            </div>
-                        </Link>
+                        <UserPlaceItem place={place} key={place._id}/>
                     ))}
                 </div>
         </div>
